// livelocalgadget3/.gadget/client/src/types.ts
interface JSONObject {
    [x: string]: JSONValue;
}
// ...
export type SortOrder = "Ascending"|"Descending";
// ...
export type BookingSort = {

  /** Sort the results by the id field. Defaults to ascending (smallest value first). */
  id?: SortOrder | null;

  /** Sort the results by the createdAt field. Defaults to ascending (smallest value first). */
  createdAt?: SortOrder | null;

  /** Sort the results by the updatedAt field. Defaults to ascending (smallest value first). */
  updatedAt?: SortOrder | null;

  /** Sort the results by the date field. Defaults to ascending (smallest value first). */
  date?: SortOrder | null;

  /** Sort the results by the depositAmount field. Defaults to ascending (smallest value first). */
  depositAmount?: SortOrder | null;

  /** Sort the results by the depositPaid field. Defaults to ascending (smallest value first). */
  depositPaid?: SortOrder | null;

  /** Sort the results by the endTime field. Defaults to ascending (smallest value first). */
  endTime?: SortOrder | null;

  /** Sort the results by the fullPaymentPaid field. Defaults to ascending (smallest value first). */
  fullPaymentPaid?: SortOrder | null;

  /** Sort the results by the isActive field. Defaults to ascending (smallest value first). */
  isActive?: SortOrder | null;

  /** Sort the results by the notes field. Defaults to ascending (smallest value first). */
  notes?: SortOrder | null;

  /** Sort the results by the specialRequirements field. Defaults to ascending (smallest value first). */
  specialRequirements?: SortOrder | null;

  /** Sort the results by the startTime field. Defaults to ascending (smallest value first). */
  startTime?: SortOrder | null;

  /** Sort the results by the status field. Defaults to ascending (smallest value first). */
  status?: SortOrder | null;

  /** Sort the results by the totalAmount field. Defaults to ascending (smallest value first). */
  totalAmount?: SortOrder | null;
};



export type BookingFilter = {

  AND?: (BookingFilter | null)[];

  OR?: (BookingFilter | null)[];

  NOT?: (BookingFilter | null)[];

  id?: IDFilter | null;

  createdAt?: DateTimeFilter | null;

  updatedAt?: DateTimeFilter | null;

  bookedById?: IDFilter | null;

  bookedBy?: UserRelationshipFilter | null;

  date?: DateTimeFilter | null;

  depositAmount?: FloatFilter | null;

  depositPaid?: BooleanFilter | null;

  endTime?: StringFilter | null;

  fullPaymentPaid?: BooleanFilter | null;

  isActive?: BooleanFilter | null;

  musicianId?: IDFilter | null;

  musician?: MusicianRelationshipFilter | null;

  notes?: StringFilter | null;

  specialRequirements?: StringFilter | null;

  startTime?: StringFilter | null;

  status?: StringFilter | null;

  totalAmount?: FloatFilter | null;

  venueId?: IDFilter | null;

  venue?: VenueRelationshipFilter | null;
};
// ...
export interface Scalars {
  /** Represents an amount of some currency. Specified as a string so user's aren't tempted to do math on the value. */
  CurrencyAmount: string;
  /** Represents a UTC date formatted an ISO-8601 formatted 'full-date' string. */
  ISO8601DateString: string;
  /** A file object produced by a browser for uploading to cloud storage */
  Upload: File;
  /** A record's current state for a recordState type field */
  StateValue: Record<string, string>;
  /** The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text. */
  String: string;
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: JSONObject;
  /** The `Boolean` scalar type represents `true` or `false`. */
  Boolean: boolean;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: JSONValue;
  /** Integer type that can handle bigints and Big numbers */
  Int: number;
  /** The ID of a record in Gadget */
  GadgetID: string;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: Date;
  /** The `Float` scalar type represents signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
  Float: number;
  /** Instructions for a client to turn raw transport types (like strings) into useful client side types (like Dates). Unstable and not intended for developer use. */
  HydrationPlan: unknown;
  /** A field whose value conforms to the standard URL format as specified in RFC3986: https://www.ietf.org/rfc/rfc3986.txt. */
  URL: string;
  /** Represents the state of one record in a Gadget database. Represented as either a string or set of strings nested in objects. */
  RecordState: (string | { [key: string]: Scalars['RecordState'] });
};
// ...
export type Query = {

  __typename: 'Query';

  session: (Session | null);

  sessions: SessionConnection;

  sessionGellyView: (Scalars['JSON'] | null);

  booking: (Booking | null);

  bookings: BookingConnection;

  bookingGellyView: (Scalars['JSON'] | null);

  event: (Event | null);

  events: EventConnection;

  eventGellyView: (Scalars['JSON'] | null);

  musician: (Musician | null);

  musicians: MusicianConnection;

  musicianGellyView: (Scalars['JSON'] | null);

  review: (Review | null);

  reviews: ReviewConnection;

  reviewGellyView: (Scalars['JSON'] | null);

  venue: (Venue | null);

  venues: VenueConnection;

  venueGellyView: (Scalars['JSON'] | null);

  user: (User | null);

  users: UserConnection;

  userGellyView: (Scalars['JSON'] | null);

  gellyView: (Scalars['JSON'] | null);

  currentSession: (Session | null);

  internal: InternalQueries;
};
// ...
export type Session = {

  __typename: 'Session';

  /** The globally unique, unchanging identifier for this record. Assigned and managed by Gadget. */
  id: (Scalars['GadgetID'] | null);

  /** The time at which this record was first created. Set once upon record creation and never changed. Managed by Gadget. */
  createdAt: Scalars['DateTime'];

  /** The time at which this record was last changed. Set each time the record is successfully acted upon by an action. Managed by Gadget. */
  updatedAt: Scalars['DateTime'];

  user: (User | null);

  userId: (Scalars['GadgetID'] | null);

  /** Get all the fields for this record. Useful for not having to list out all the fields you want to retrieve, but slower. */
  _all: Scalars['JSONObject'];
};
// ...
export type SessionConnection = {

  __typename: 'SessionConnection';

  /** A list of edges. */
  edges: SessionEdge[];

  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type Booking = {

  __typename: 'Booking';

  /** The globally unique, unchanging identifier for this record. Assigned and managed by Gadget. */
  id: Scalars['GadgetID'];

  /** The time at which this record was first created. Set once upon record creation and never changed. Managed by Gadget. */
  createdAt: Scalars['DateTime'];

  /** The time at which this record was last changed. Set each time the record is successfully acted upon by an action. Managed by Gadget. */
  updatedAt: Scalars['DateTime'];

  bookedBy: (User | null);

  bookedById: (Scalars['GadgetID'] | null);

  date: (Scalars['DateTime'] | null);

  depositAmount: (Scalars['Float'] | null);

  depositPaid: (Scalars['Boolean'] | null);

  endTime: (Scalars['String'] | null);

  fullPaymentPaid: (Scalars['Boolean'] | null);

  isActive: (Scalars['Boolean'] | null);

  musician: (Musician | null);

  musicianId: (Scalars['GadgetID'] | null);

  notes: (Scalars['String'] | null);

  specialRequirements: (Scalars['String'] | null);

  startTime: (Scalars['String'] | null);

  status: (Scalars['String'] | null);

  totalAmount: (Scalars['Float'] | null);

  venue: (Venue | null);

  venueId: (Scalars['GadgetID'] | null);

  /** Get all the fields for this record. Useful for not having to list out all the fields you want to retrieve, but slower. */
  _all: Scalars['JSONObject'];
};



export type AvailableBookingSelection = {

  __typename?: boolean | null | undefined;

  /** The globally unique, unchanging identifier for this record. Assigned and managed by Gadget. */
  id?: boolean | null | undefined;

  /** The time at which this record was first created. Set once upon record creation and never changed. Managed by Gadget. */
  createdAt?: boolean | null | undefined;

  /** The time at which this record was last changed. Set each time the record is successfully acted upon by an action. Managed by Gadget. */
  updatedAt?: boolean | null | undefined;

  bookedBy?: AvailableUserSelection;

  bookedById?: boolean | null | undefined;

  date?: boolean | null | undefined;

  depositAmount?: boolean | null | undefined;

  depositPaid?: boolean | null | undefined;

  endTime?: boolean | null | undefined;

  fullPaymentPaid?: boolean | null | undefined;

  isActive?: boolean | null | undefined;

  musician?: AvailableMusicianSelection;

  musicianId?: boolean | null | undefined;

  notes?: boolean | null | undefined;

  specialRequirements?: boolean | null | undefined;

  startTime?: boolean | null | undefined;

  status?: boolean | null | undefined;

  totalAmount?: boolean | null | undefined;

  venue?: AvailableVenueSelection;

  venueId?: boolean | null | undefined;

  /** Get all the fields for this record. Useful for not having to list out all the fields you want to retrieve, but slower. */
  _all?: boolean | null | undefined;
};
// ...
export type EventConnection = {

  __typename: 'EventConnection';

  /** A list of edges. */
  edges: EventEdge[];

  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type Event = {

  __typename: 'Event';

  /** The globally unique, unchanging identifier for this record. Assigned and managed by Gadget. */
  id: Scalars['GadgetID'];

  /** The time at which this record was first created. Set once upon record creation and never changed. Managed by Gadget. */
  createdAt: Scalars['DateTime'];

  /** The time at which this record was last changed. Set each time the record is successfully acted upon by an action. Managed by Gadget. */
  updatedAt: Scalars['DateTime'];

  availableTickets: (Scalars['Float'] | null);

  category: (Scalars['String'] | null);

  createdBy: (User | null);

  createdById: (Scalars['GadgetID'] | null);

  date: (Scalars['DateTime'] | null);

  description: (Scalars['String'] | null);

  endTime: (Scalars['String'] | null);

  image: (Scalars['String'] | null);

  isActive: (Scalars['Boolean'] | null);

  isPublic: (Scalars['Boolean'] | null);

  musician: (Musician | null);

  musicianId: (Scalars['GadgetID'] | null);

  setlist: (Scalars['JSON'] | null);

  startTime: (Scalars['String'] | null);

  status: (Scalars['String'] | null);

  ticketPrice: (Scalars['Float'] | null);

  ticketType: (Scalars['String'] | null);

  title: (Scalars['String'] | null);

  totalCapacity: (Scalars['Float'] | null);

  venue: (Venue | null);

  venueId: (Scalars['GadgetID'] | null);

  /** Get all the fields for this record. Useful for not having to list out all the fields you want to retrieve, but slower. */
  _all: Scalars['JSONObject'];
};
// ...
export type BookingConnection = {

  __typename: 'BookingConnection';

  /** A list of edges. */
  edges: BookingEdge[];

  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type MusicianConnection = {

  __typename: 'MusicianConnection';

  /** A list of edges. */
  edges: MusicianEdge[];

  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type PromiseOrLiveIterator<Options extends { live?: boolean }, Result> = Options extends { live: true } ? AsyncIterable<Result> : Promise<Result>;